+++ src/yaml_include/yaml_types_backward.py
@@ -1,7 +1,7 @@
 import sys
 
 if not sys.version_info < (3, 12):
-    raise ImportError("Python greater than or equal to 3.12 should “import yamltypes”")
+    raise ImportError("Python version greater than or equal to 3.12 should “from . import yaml_types”")
 
 from typing import Type, Union